+++ include/Network/Server.h
@@ -52,7 +52,7 @@ public:
 	void sendEdgeColor(int id, float r, float g, float b, float alpha, QTcpSocket *client = NULL);
 	void sendNodeLabel(int id, QString label, QTcpSocket *client = NULL);
 	void sendFixNodeState(int id, bool state, QTcpSocket *client = NULL);
-	void sendMergeNodes(QLinkedList<osg::ref_ptr<Data::Node> > * selectedNodes, osg::Vec3f position, int mergeNodeId, QTcpSocket *client = NULL);
+	void sendMergeNodes(QLinkedList<osg::ref_ptr<Data::Node> > * selectedNodes, osg::Vec3f position, qlonglong mergeNodeId, QTcpSocket *client = NULL);
 	void sendSeparateNodes(QLinkedList<osg::ref_ptr<Data::Node> > * selectedNodes, QTcpSocket *client = NULL);
 	void sendAddMetaNode(osg::ref_ptr<Data::Node> metaNode, QLinkedList<osg::ref_ptr<Data::Node> > * selectedNodes, QString edgeName, osg::Vec3f position, QTcpSocket *client = NULL);
 	void sendSetRestriction(quint8 type, osg::ref_ptr<Data::Node> node1, osg::Vec3 position_node1,  osg::ref_ptr<Data::Node> node2, osg::Vec3 position_node2, QLinkedList<osg::ref_ptr<Data::Node> > * nodes, osg::ref_ptr<Data::Node> node3 = NULL, osg::Vec3 * position_node3 = NULL, QTcpSocket *client = NULL);
