+++ src/Layout/ShapeVisitor_RestrictedPositionGetter.cpp
@@ -181,7 +181,7 @@ osg::Vec3f ShapeVisitor_RestrictedPositionGetter::toSphere(const osg::Vec3f &cen
 	if (changedPointMoved.length() == 0.0) {
 		// create random point
 		for (int i = 0; i < 3; ++i) {
-			changedPointMoved[i] = (rand () % 100) + 1;
+			changedPointMoved[i] = (float)(rand () % 100) + 1.f;
 		}
 	}
