+++ src/Math/DjikstraAlg.cpp
@@ -39,8 +39,8 @@ void DjikstraAlg::findMinimalDistances(Data::Node * node)
 	{
 		Data::Node * target = adjacentNodes.at(x);
 
-		int tSh = getShortestDistance(target);
-		int nSh = getShortestDistance(node);
+		qlonglong tSh = getShortestDistance(target);
+		qlonglong nSh = getShortestDistance(node);
 		int d = getDistance(node, target);
 
 		if ( tSh > nSh + d)
@@ -127,7 +127,7 @@ bool DjikstraAlg::isSettled(Data::Node * vertex)
 	return settledNodes.contains(vertex->getId());
 }
 
-int DjikstraAlg::getShortestDistance(Data::Node * destination)
+qlonglong DjikstraAlg::getShortestDistance(Data::Node * destination)
 {
 	qlonglong d = -1;
